backend/requirements.txt
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
opik
attrs==25.4.0
bidict==0.23.1
black==25.11.0
blinker==1.9.0
cachetools==6.2.2
cairosvg==2.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.24.0
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fonttools==4.61.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.52.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
openai==2.8.1
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
playwright==1.56.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.4.0
pytest==9.0.1
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-engineio==4.12.3
python-pptx==1.0.2
python-socketio==5.15.0
psycopg2-binary==2.9.10
pymysql==1.1.2
cryptography==42.0.8
pytokens==0.3.0
pytz==2025.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
supabase==2.15.0
tavily-python==0.7.13
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
mcp>=1.0.0
websockets>=13.0
Werkzeug==3.1.3
gunicorn==23.0.0
gevent==24.11.1
gevent-websocket==0.10.1
wsproto==1.3.2
xlsxwriter==3.2.9
youtube-transcript-api==1.2.3
